Adding `Elem2x16` SIMD type; unfortunate allocations.
Added `Elem2x16` SIMD type along with `warmup.jl` to ensure that the
compiler gets a little help. Unfortunately, it seems like there are
allocations when performing `fft_mul!` with this type that I cannot
currently avoid. Might be worth looking into later. (One idea is to
reinterpret as `UInt32`s and then operate on that, which perhaps the
compiler might be happier with. In the worst case, we just write SIMD
code directly for this special case.)